+++ src/pages/SaveEditor/pages/General/selectors.ts
@@ -1,0 +1,19 @@
+
+import { createStructuredSelector } from "@/state/utils";
+
+import {
+    cycles,
+    nextSpawn,
+    isSpawnReady
+} from "@/services/save-editor/general/selectors";
+
+export interface StateProps {
+    cycles: number;
+    nextSpawn: number;
+    isSpawnReady: boolean;
+}
+export default createStructuredSelector<StateProps>({
+    cycles,
+    nextSpawn,
+    isSpawnReady
+});