+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+### v1.3.0
+- Edit current cycle.
+- Edit duplicant printer
+    - Next duplicant ready
+    - Time to next duplicant
+
 ### v1.2.1
 - Fix save corruption when text with multi-byte accents are encountered.
 - Refactor URL layout to make way for future utilities.